Time Tracker
Time tracker shows under the map, its not completely functional, but in
theory will track the time consumed by actions on the dice. Also, when
the player selects an action, it will light up yellow showing how much
that action costs. After it reaches the end, a provision should be
consumed.